much better TS commands thanks to declare global
