+++ cypress/plugins/index.js
@@ -1,27 +1,5 @@
-const wp = require('@cypress/webpack-preprocessor')
-
-const webpackOptions = {
-  resolve: {
-    extensions: ['.ts', '.js']
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        exclude: [/node_modules/],
-        use: [
-          {
-            loader: 'ts-loader'
-          }
-        ]
-      }
-    ]
-  }
-}
+const cypressTypeScriptPreprocessor = require('./cy-ts-preprocessor')
 
 module.exports = on => {
-  const options = {
-    webpackOptions
-  }
-  on('file:preprocessor', wp(options))
+  on('file:preprocessor', cypressTypeScriptPreprocessor)
 }
