--- cypress/support/actions.ts
@@ -1,22 +1,0 @@
-/**
- * An example function "foo()"
- *
- * @returns {string} "foo"
- * @example
- *    foo() // "foo"
- */
-export function foo() {
-  return 'foo'
-}
-
-/**
- * Uses cy.foo() internally
- *
- * @returns
- */
-export function foo2() {
-  return cy.foo()
-}
-
-const name = 'actions'
-console.log('name in actions is', name)
