+++ package.json
@@ -13,9 +13,9 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@bahmutov/add-typescript-to-cypress": "^1.1.1",
-    "@types/node": "^8.5.5",
-    "cypress": "^2.1.0",
-    "typescript": "^2.8.3"
+    "@bahmutov/add-typescript-to-cypress": "2.0.0",
+    "@types/node": "8.10.36",
+    "cypress": "3.1.0",
+    "typescript": "3.1.3"
   }
 }
